requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pgvector==0.2.5
celery==5.3.6
redis==5.0.1
groq>=0.9.0
httpx>=0.27.0
langchain==0.1.9
langchain-community==0.0.24
langchain-groq==0.0.1
sentence-transformers==2.5.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
requests==2.31.0
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1